# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ToolsHub API - Tool Rental Management Platform",
    "version": "1.0.0",
    "description": "API documentation for the ToolsHub application with CRUD operations for tools and rentals."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Rental": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "toolId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Completed",
              "Cancelled",
              "Overdue"
            ]
          },
          "lateFee": {
            "type": "number"
          },
          "reminderSent": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "returnedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateRentalDTO": {
        "type": "object",
        "required": [
          "toolId",
          "customerId",
          "quantity",
          "startDate",
          "endDate"
        ],
        "properties": {
          "toolId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/rentals/customer/{customerId}": {
      "get": {
        "summary": "Get rentals by customer",
        "description": "Retrieve all rental contracts for a specific customer",
        "tags": [
          "Rentals"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Rentals retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Rental"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rentals": {
      "get": {
        "summary": "Get all rentals",
        "description": "Retrieve a list of all rental contracts",
        "tags": [
          "Rentals"
        ],
        "responses": {
          "200": {
            "description": "Rentals retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Rental"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new rental",
        "description": "Create a new rental contract for a tool",
        "tags": [
          "Rentals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentalDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rental created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Rental"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tool not available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rentals/{id}": {
      "get": {
        "summary": "Get rental by ID",
        "description": "Retrieve a single rental contract by its unique identifier",
        "tags": [
          "Rentals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Rental ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Rental retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Rental"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Rental not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rentals/{id}/cancel": {
      "post": {
        "summary": "Cancel a rental",
        "description": "Cancel an active rental contract",
        "tags": [
          "Rentals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Rental ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Rental cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Rental"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Rental not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}